+++ CHANGELOG.md
@@ -104,3 +104,4 @@ ADD ALL KBO DATA
 
     1.16.16
     fix: small copy improvement for adres
+    fix: ignore empty contact points